Replace all parted calls with parted --script
In some configurations, calls to `parted ...` will render a user prompt
(eg. when a partition table on a disk is unusable, such as for some
bootable usb keys created by unetbootin). In this case, boot-repair will
hang waiting for the partition scan to complete, and nothing is printed
to the terminal. This patch enables the parted script flag for every
parted call which will never present this kind of interactive prompt.